+++ app/src/main/kotlin/io/github/bsfranca2/athena/controller/IssueController.kt
@@ -20,27 +20,27 @@ class IssueController(val issueService: IssueService) {
             = issueService.listIssues()
 
     @PutMapping("/{id}")
-    fun updateIssue(@PathVariable id: Int, @Valid @RequestBody issueUpdateDto: RequestIssueDto)
+    fun updateIssue(@PathVariable id: Long, @Valid @RequestBody issueUpdateDto: RequestIssueDto)
             = issueService.updateIssue(id, issueUpdateDto)
 
     @DeleteMapping("/{id}")
-    fun deleteIssue(@PathVariable id: Int)
+    fun deleteIssue(@PathVariable id: Long)
             = issueService.deleteIssue(id)
 
     @PostMapping("/{issueId}/time-entries") @ResponseStatus(HttpStatus.CREATED)
-    fun addTimeEntry(@PathVariable issueId: Int, @RequestBody timeEntryDto: TimeEntryDto)
+    fun addTimeEntry(@PathVariable issueId: Long, @RequestBody timeEntryDto: TimeEntryDto)
             = issueService.addTimeEntry(issueId, timeEntryDto)
 
     @GetMapping("/{issueId}/time-entries")
-    fun listTimeEntries(@PathVariable issueId: Int)
+    fun listTimeEntries(@PathVariable issueId: Long)
             = issueService.listTimeEntries(issueId)
 
     @PutMapping("/{issueId}/time-entries/{id}")
-    fun updateTimeEntry(@PathVariable issueId: Int, @PathVariable id: Int, @Valid @RequestBody timeEntryDto: TimeEntryDto)
+    fun updateTimeEntry(@PathVariable issueId: Long, @PathVariable id: Long, @Valid @RequestBody timeEntryDto: TimeEntryDto)
             = issueService.updateTimeEntry(issueId, id, timeEntryDto)
 
     @DeleteMapping("/{issueId}/time-entries/{id}")
-    fun removeTimeEntry(@PathVariable issueId: Int, @PathVariable id: Int)
+    fun removeTimeEntry(@PathVariable issueId: Long, @PathVariable id: Long)
             = issueService.removeTimeEntry(issueId, id)
 
 }
