+++ app/src/main/kotlin/io/github/bsfranca2/athena/dto/issue/StoryIssueDto.kt
@@ -4,7 +4,7 @@ import io.github.bsfranca2.athena.enum.IssueType
 import java.time.LocalDateTime
 
 class StoryIssueDto(
-        id: Int,
+        id: Long,
         title: String,
         description: String,
         status: String,
@@ -12,8 +12,8 @@ class StoryIssueDto(
         startDate: LocalDateTime?,
         endDate: LocalDateTime?,
         val storyPoints: Int = 0,
-        val parentId: Int? = null,
+        val parentId: Long? = null,
         val children: MutableList<IssueDto> = mutableListOf(),
-        val assignedTo: MutableList<Int> = mutableListOf(),
-        createdBy: Int
+        val assignedTo: MutableList<Long> = mutableListOf(),
+        createdBy: Long
 ) : IssueDto(id, IssueType.STORY, title, description, status, priority, startDate, endDate, createdBy)
