+++ app/src/main/kotlin/io/github/bsfranca2/athena/exception/UserNotFoundException.kt
@@ -1,3 +1,3 @@
 package io.github.bsfranca2.athena.exception
 
-data class UserNotFoundException(val id: Int) : Throwable("User $id not found")
+data class UserNotFoundException(val id: Long) : Throwable("User $id not found")
