+++ app/src/main/kotlin/io/github/bsfranca2/athena/repository/UserRepository.kt
@@ -3,6 +3,6 @@ package io.github.bsfranca2.athena.repository
 import io.github.bsfranca2.athena.entity.User
 import org.springframework.data.repository.CrudRepository
 
-interface UserRepository : CrudRepository<User, Int> {
+interface UserRepository : CrudRepository<User, Long> {
     fun findByEmail(email: String?): User?
 }
